pyproject.toml
[project]
name = "spectral-web"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.122.0",
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "pillow>=12.0.0",
    "setuptools>=80.9.0",
    "spectral>=0.24",
    "uvicorn>=0.38.0",
]
